phase-5/services/recurring-task-service/requirements.txt
# Recurring Task Service Dependencies
fastapi>=0.124.0
uvicorn[standard]>=0.38.0
dapr-ext-fastapi>=1.15.0
aiokafka>=0.11.0
python-dateutil>=2.9.0
sqlalchemy[asyncio]>=2.0.45
asyncpg>=0.31.0
pydantic>=2.12.5
pydantic-settings>=2.12.0
python-dotenv>=1.2.1
structlog>=24.1.0
pytest>=9.0.2
pytest-asyncio>=1.3.0
pytest-cov>=7.0.0
httpx>=0.28.1
